Updated AuthenticationResponse authenticate return

+++ src/main/java/com/example/securityumarket/dao/AppUserDAO.java
@@ -1,10 +1,10 @@
     package com.example.securityumarket.dao;
 
-    import com.example.securityumarket.models.AppUser;
+    import com.example.securityumarket.models.entities.AppUser;
     import org.springframework.data.jpa.repository.JpaRepository;
 
     import java.util.Optional;
 
-    public interface AppUserDAO extends JpaRepository<AppUser, Integer> {
+    public interface AppUserDAO extends JpaRepository<AppUser, Long> {
         Optional<AppUser> findAppUserByEmail(String email);
     }
